added geometry validation semantics
