+++ engine/runtime/geometry/src/solid/validation.rs
@@ -1,6 +1,6 @@
 use super::Solid;
-use crate::ops::validation::{check_finite_3d, Validate, ValidationReport};
-use crate::ops::ValidationType;
+use crate::validation_next::ValidationType;
+use crate::validation_next::{check_finite_3d, Validate, ValidationReport};
 
 impl Validate for Solid {
     fn validate(&self, _valid_type: ValidationType) -> Option<ValidationReport> {
